% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sens_multi.R
\name{find_odds_multi}
\alias{find_odds_multi}
\title{Calculate the difference between the p-value and a p-value threshold for the multiple types of information model}
\usage{
find_odds_multi(omega, obs_support, rival_obs, weights, p_threshold)
}
\arguments{
\item{omega}{Candidate odds ratio being evaluated by the root finder.}

\item{obs_support}{A vector of integers representing the number of
observations made in favor of the working hypothesis (anti-rival).
Each element corresponds to evidence against a specific rival.}

\item{rival_obs}{Optional. A vector of integers representing the number of
observations actually made that support each rival. Must be the same length
as obs_support. If NULL, assumes zero pro-rival observations.}

\item{weights}{Not used yet.}

\item{p_threshold}{A number between 0 and 1 indicating the critical value of the test.}
}
\description{
A helper function for \code{sens_obs_two_types()} see \code{find_p_multi_mv()} for details.
}
