% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sens_binary.R
\name{find_odds_two_types}
\alias{find_odds_two_types}
\title{Calculate the difference between the p-value and a p-value threshold for the two types of observation model}
\usage{
find_odds_two_types(
  omega,
  obs_support,
  total_obs,
  rival_obs,
  weights,
  p_threshold
)
}
\arguments{
\item{obs_support}{An integer representing the number of observations in
favor of the working hypothesis. Must be less than or equal to \code{total_obs}.}

\item{total_obs}{An integer representing the total number of observations
made. It can be greater than or equal to the \code{obs_support}.}

\item{rival_obs}{Optional. An integer representing the number of
observations in the urn that do not support the working theory.}

\item{weights}{Double vector. Default is equal weight for each observation when
weights=NULL is \code{rep(1,obs_support)}. To indicate that one observation
should have twice the weight of any other one might use
\code{rep(c(2,1),c(1,obs_support-1))}}

\item{p_threshold}{A number between 0 and 1 indicating the critical value of the test.}
}
\description{
A helper function for \code{sens_obs_two_types()} see \code{find_p_two_types()} for details.
}
