% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_multiple_mv.R
\name{generate_compositions}
\alias{generate_compositions}
\title{Generate all compositions of n into k parts}
\usage{
generate_compositions(n, k)
}
\arguments{
\item{n}{Integer to partition}

\item{k}{Number of parts}
}
\value{
Matrix where each row is a k-way composition of n
}
\description{
This is a helper function for \code{find_p_multi_mv()}
}
