% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_odds.R
\name{sens_analysis}
\alias{sens_analysis}
\title{Find odds given ideas about unequally easy evidence}
\usage{
sens_analysis(obs_support, obs_oppose, total_obs, thep = 0.05)
}
\arguments{
\item{obs_support}{An integer representing the number of observations in favor of the working hypothesis. Must be less than or equal to the total.}

\item{total_obs}{An integer representing the total number of observations}

\item{thep}{The p-value threshold}
}
\value{
The odds required for the p-value to be equal to that specified in thep
}
\description{
Find odds
}
\details{
A function that allows the researcher to input the likelihood of data or to find how mush bias in datacollection would be nescessary to obtain a p>0.05 or p>0.10 result using the same data.
}
