% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sens_multi.R
\name{sens_obs_multi}
\alias{sens_obs_multi}
\title{Sensitivity analysis for urn models with two types of information}
\usage{
sens_obs_multi(
  obs_support,
  rival_obs = NULL,
  weights = NULL,
  p_threshold = 0.05
)
}
\arguments{
\item{obs_support}{An integer representing the number of observations
in favor of the working hypothesis. Must be less than or equal to the total.}

\item{rival_obs}{Optional. The number of observations in the urn that do not support the working theory.}

\item{weights}{A vector of numeric weights representing the differential
evidentiary weight of the working theory supporting observations.}

\item{p_threshold}{A decimal (Double). The p-value threshold. Default is p=.05.}
}
\value{
A list with two elements:
\enumerate{
\item in \code{w} the differential odds of observing a working theory
supporting observation over a rival theory supporting observation.
\item in \code{p} The p-value given the found \code{w}.
}
}
\description{
A function that allows the researcher to calculate the odds of
drawing a working theory supporting observation over a rival theory supporting observation
required to increase a p-value to above a certain value.
}
\examples{
# What is the odds that would bring our p=.02 to p \approx .05
find_p_multi_mv(obs_support = c(4, 3, 2, 1), odds = 1)
find_p_multi_mv(obs_support = c(4, 3, 2, 1), odds = 2)
sens_obs_multi(obs_support = c(4, 3, 2, 1), p_threshold = .05)$w
# Notice that this is correct:
find_p_multi_mv(obs_support = c(4, 3, 2, 1), odds = 22.535)
sens_obs_multi(obs_support = c(4, 3, 2, 1), rival_obs = c(1, 1, 0, 0), p_threshold = .05)$w
sens_obs_multi(obs_support = c(4, 4, 4, 4), p_threshold = .05)
find_p_multi_mv(obs_support = c(4, 4, 4, 4), odds = 2.6)
}
