% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_analysis.R
\name{sens_urn}
\alias{sens_urn}
\title{Sensitivity analysis for urn models with two types of information}
\usage{
sens_urn(obs_support, total_obs, weights = NULL, p_threshold = 0.05)
}
\arguments{
\item{obs_support}{An integer representing the number of observations
in favor of the working hypothesis. Must be less than or equal to the total.}

\item{total_obs}{An integer representing the total number of observations}

\item{weights}{A vector of numeric weights representing the differential evidentiary weight of the working theory supporting observations.}

\item{p_threshold}{A decimal (Double). The p-value threshold. Default is p=.05.}
}
\value{
The p-value for a given amount of bias in drawing observations from the urn.
}
\description{
A function that allows the researcher to calculate the odds of
drawing a working theory supporting observation over a rival theory supporting observation
required to increase a p-value to above a certain value.
}
