% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_urn.R
\name{sim_urn}
\alias{sim_urn}
\title{Simulate draws from an urn model}
\usage{
sim_urn(urn_size, total_obs, item_probs, thep)
}
\arguments{
\item{urn_size}{Total number of items in the urn}

\item{total_obs}{An integer representing the total number of observations}

\item{thep}{The p-value threshold}
}
\value{
A vector of length n
}
\description{
Simulate an urn model
}
\details{
A function that allows the researcher to simulate draws from urn models that
do not have nice analytical expressions. For example, urn models where each
ball has a different probability of being drawn. Given a set of draws, we
can calculate probabilities.
}
