% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_odds.R
\name{sens_analysis}
\alias{sens_analysis}
\title{Find odds given ideas about unequally easy evidence}
\usage{
sens_analysis(m1, m2, n, thep)
}
\arguments{
\item{m1}{TODO}

\item{m2}{TODO}

\item{n}{TODO}

\item{thep}{TODO}
}
\value{
A vector representing TODO
}
\description{
Find odds
}
\details{
TODO Description
}
