% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_odds.R
\name{sens_urn}
\alias{sens_urn}
\title{Sensitivity analysis for simple urn models}
\usage{
sens_urn(obs_support, obs_oppose, p_threshold = 0.05)
}
\arguments{
\item{obs_support}{An integer representing the number of observations
in favor of the working hypothesis. Must be less than or equal to the total.}

\item{obs_oppose}{An integer representing the number of observations in against
the working hypothesis. Must be less than or equal to the total.}

\item{total_obs}{An integer representing the total number of observations}

\item{thep}{A decimal (Double). The p-value threshold. Default is p=.05.}
}
\value{
The p-value for a given amount of bias in drawing observations from the urn.
}
\description{
A function that allows the researcher to calculate the p-value arising from an urn
where the probability of drawing a rival supporting observation is not equal to the
probability of drawing a working-theory supporting observation.
}
