% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_analysis.R
\name{sens_urn}
\alias{sens_urn}
\title{Sensitivity analysis for urn models with two types of information}
\usage{
sens_urn(obs_support, total_obs, weights = NULL, p_threshold = 0.05)
}
\arguments{
\item{obs_support}{An integer representing the number of observations
in favor of the working hypothesis. Must be less than or equal to the total.}

\item{total_obs}{An integer representing the total number of observations}

\item{weights}{A vector of numeric weights representing the differential evidentiary weight of the working theory supporting observations.}

\item{p_threshold}{A decimal (Double). The p-value threshold. Default is p=.05.}
}
\value{
A list with two elements:
(1) in \code{w} the differential odds of observing a working theory supporting observation over a rival theory supporting observation.
(2) in \code{p} The p-value given the found \code{w}.
}
\description{
A function that allows the researcher to calculate the odds of
drawing a working theory supporting observation over a rival theory supporting observation
required to increase a p-value to above a certain value.
}
\examples{
# What is the odds that would bring our p=.02 to p \approx .05
find_p_two_types(obs_support = 7, total_obs = 10, odds = 1)
sens_urn(obs_support = 7, total_obs = 10, p_threshold = .05)$w
# Notice that this is correct:
find_p_two_types(obs_support = 7, total_obs = 10, odds = 1.58928)
}
